# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Yamen Business Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchaseOrders
      tags: [purchases]
      summary: List purchase orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Purchase orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchases]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchases/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchases]
      summary: Get purchase order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    put:
      operationId: updatePurchaseOrder
      tags: [purchases]
      summary: Update purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchases]
      summary: Delete purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [id, productId, productName, quantity, unitPrice, total]
...
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
...
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, received, cancelled]
        totalAmount:
          type: number
        paidAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNumber, status, totalAmount, paidAmount, createdAt]
    CreatePurchaseOrderBody:
      type: object
      properties:
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
        paidAmount:
          type: number
        notes:
          type: string
          nullable: true
      required: [items, paidAmount]
    UpdatePurchaseOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, received, cancelled]
        paidAmount:
          type: number
      required: [status]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        unitPrice:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: number
        minStockLevel:
          type: number
        unit:
          type: string
        createdAt:
          type: string
      required: [id, name, unitPrice, costPrice, stockQuantity, minStockLevel, unit, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        unitPrice:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: number
        minStockLevel:
          type: number
        unit:
          type: string
      required: [name, unitPrice, costPrice, stockQuantity, minStockLevel, unit]
...